pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "LangGraph Platform-compatible conversational AI agent using OpenAI GPT-4o-mini with full API support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.27",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.8.0",
    "langchain-tavily>=0.2.11",
    "langchain-mcp-adapters>=0.1.9",
    "simple-salesforce>=1.12.9",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Test files are now organized in the test/ directory
# Only main.py module will be packaged for deployment
py-modules = ["main"]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=7.0.0",
]
